backend/requirements.txt
fastapi
uvicorn
python-socketio
opencv-python
mediapipe
librosa
soundfile
openai
textblob
supabase
python-dotenv
google-generativeai
numpy
requests
